Fix parsing of <double> in new XMLRPC code
The new XMLRPC code didn't parse <double>s correctly, which we didn't
notice because the test cases don't test <double> either. So fix the
code and the tests.

https://bugzilla.gnome.org/show_bug.cgi?id=767707

+++ tests/xmlrpc-server.php
@@ -18,7 +18,7 @@ function sum ($method_name, $params, $app_data)
 
 	$sum = 0;
 	foreach ($params[0] as $val) {
-		if (xmlrpc_get_type ($val) != "int")
+		if (xmlrpc_get_type ($val) != "double")
 			return paramfault();
 			
 		$sum = $sum + $val;
